### 2022-01-12 #499 @sandrinecharles
`rbioacc`: an `R`-package to analyze toxicokinetic data
**Reviewers:**
Submitting Author Name: Sandrine Charles
Submitting Author Github Handle: @sandrinecharles
Other Package Authors Github handles: (comma separated, delete if none) @aursiber, @virgile-baudrot
Repository:  https://github.com/aursiber/rbioacc
Version submitted:
Submission type: Standard

Version accepted: TBD
---

```
Package: rbioacc
Title: Inference and Prediction of ToxicoKinetic (TK) Models
Version: 1.1-0
    person("Virgile", "Baudrot", role = "aut", email = "virgile.baudrot@posteo.net"),
    person("Sandrine", "Charles", role = "aut"),
    person("Ophélia", "Gestin", role = "ctb"),
    person("Mélina", "Kaag", role = "aut"),
    person("Christelle", "Lopes", role = "ctb"),
    person("Gauthier", "Multari", role = "ctb"),
    person("Alain", "Pavé", role = "ctb"),
    person("Aude", "Ratier", role = "aut"),
    person("Aurélie", "Siberchicot", role = c("aut", "cre"), email = "aurelie.siberchicot@univ-lyon1.fr")
    )
Description: The MOSAICbioacc application is a turnkey package providing bioaccumulation
    factors (BCF/BMF/BSAF) from a toxicokinetic (TK) model fitted to
    accumulation-depuration data. It is designed to fulfil the requirements
    of regulators when examining applications for market authorization of active
    substances. See Ratier et al. (2021) <doi:10.1101/2021.09.08.459421>.
Biarch: true
    Rcpp,
    rstan (>= 2.18.1),
    rstantools (>= 2.1.1),
    ggmcmc,
    GGally,
    loo,
    zoo
    BH (>= 1.66.0),
    Rcpp,
    RcppEigen (>= 0.3.3.3.0),
    RcppParallel (>= 5.0.1),
    rstan (>= 2.18.1),
    StanHeaders (>= 2.18.0)
SystemRequirements: GNU make

```

The `rbioacc` package provides ready-to-use functions to fit and predict accumulation and depuration of chemical substances within organisms, based on one-compartment toxicokinetics models. This package supports the freely available on-line web platform MOSAIC_bioacc (https://mosaic.univ-lyon1.fr/bioacc).

The target audience is academic research, regulatory bodies and chemical companies. Indeed, this package provide all necessary tools to estimate bioaccumulation capacity of a any chemical substance when a given living species is exposed to.

To our knowledge, no other package with the same feature exists.

Yes
