### 2022-05-12 #499 @emilyriederer
Hi there, @sandrinecharles ! I wanted to follow up on this issue. Were you still interested in addressing some of @ldecicco-USGS 's feedback and moving towards the review process? I added the "holding" tag since there hasn't been much recent discussion on this issue, but happy to continue the conversation as you wish. Thanks!
