### 2022-07-18 3cee28e
Update README with badges and doi link
